SP500 Predictor/requirements.txt
alpaca-py>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
streamlit>=1.30.0
joblib>=1.3.0
ta>=0.10.0
matplotlib>=3.7.0
plotly>=5.15.0
azure-storage-blob>=12.0.0
azure-data-tables>=12.4.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
pyyaml>=6.0
fredapi>=0.5.1
google-genai>=0.1.0
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
cryptography>=41.0.0
huggingface_hub
lxml>=5.0.0
supabase>=2.0.0
yfinance>=0.2.0
tiingo>=0.5.0

# --- F1 Predictive Modeling ---
fastf1>=3.3.0
jupyter>=1.0.0
seaborn>=0.13.0
statsmodels>=0.14.0

# --- Decoupled API Layer ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0

# --- Sports & Crypto Engines ---
websockets>=12.0
python-telegram-bot>=21.0
